multiple product images

+++ C_C_Project_2024/Areas/Admin/Controllers/ProductController.cs
@@ -46,37 +46,16 @@ namespace C_C_Proj_WebStore.Areas.Admin.Controllers
             else
             {
                 //update
-                productVM.Product = _unitOfWork.Product.Get(u=>u.Id == id);
+                productVM.Product = _unitOfWork.Product.Get(u=>u.Id == id,includeProperties: "ProductImages");
                 return View(productVM);
             }
         }
         [HttpPost]
-        public IActionResult Upsert(ProductVM productVM, IFormFile? file)
+        public IActionResult Upsert(ProductVM productVM, List<IFormFile> files)
         {
             if (ModelState.IsValid)
             {
-                string wwwRootPath = _webHostEnvironment.WebRootPath;
-                if(file != null)
-                {
-                    string fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
-                    string productPath = Path.Combine(wwwRootPath, @"Images\Product");
 
-                    if(!string.IsNullOrEmpty(productVM.Product.ImageUrl))
-                    {
-                        //delete the old image
-                        var oldImagePath = Path.Combine(wwwRootPath, productVM.Product.ImageUrl.TrimStart('\\'));
-                        if(System.IO.File.Exists(oldImagePath))
-                        {
-                            System.IO.File.Delete(oldImagePath);
-                        }
-                    }
-
-                    using (var fileStream = new FileStream(Path.Combine(productPath, fileName), FileMode.Create))
-                    {
-                        file.CopyTo(fileStream);
-                    }
-                    productVM.Product.ImageUrl = @"\Images\Product\" + fileName;
-                }
                 if (productVM.Product.Id != 0)
                 {
                     _unitOfWork.Product.Update(productVM.Product);
@@ -86,7 +65,45 @@ namespace C_C_Proj_WebStore.Areas.Admin.Controllers
                     _unitOfWork.Product.Add(productVM.Product);
                 }
                 _unitOfWork.Save();
-                TempData["success"] = "Product has been created successfully";
+
+
+                string wwwRootPath = _webHostEnvironment.WebRootPath;
+                if(files != null)
+                {
+
+                    foreach(IFormFile file in files)
+                    {
+                        string fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
+                        string productPath = @"image\products\product-" + productVM.Product.Id;
+                        string finalPath = Path.Combine(wwwRootPath, productPath);
+
+                        if(!Directory.Exists(finalPath))
+                            Directory.CreateDirectory(finalPath);
+
+                        using (var fileStream = new FileStream(Path.Combine(finalPath, fileName), FileMode.Create))
+                        {
+                            file.CopyTo(fileStream);
+                        }
+
+                        ProductImage productImage = new()
+                        {
+                            ImageUrl = @"\" + productPath + @"\" + fileName,
+                            ProductId = productVM.Product.Id
+                        };
+
+                        if(productVM.Product.ProductImages == null)
+                            productVM.Product.ProductImages = new List<ProductImage>();
+
+                        productVM.Product.ProductImages.Add(productImage);
+                        _unitOfWork.ProductImage.Add(productImage);
+                        _unitOfWork.Save();
+                    }
+
+                    //_unitOfWork.Product.Update(productVM.Product);
+                    //_unitOfWork.Save();
+                }
+                
+                TempData["success"] = "Product has been created/updated successfully";
                 return RedirectToAction("Index");
             }
             else
@@ -124,15 +141,58 @@ namespace C_C_Proj_WebStore.Areas.Admin.Controllers
             {
                 return Json(new { success = false, message = "Error while deleting" });
             }
-            var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, productToBeDeleted.ImageUrl.TrimStart('\\'));
-            if (System.IO.File.Exists(oldImagePath))
+            /* var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, productToBeDeleted.ImageUrl.TrimStart('\\'));*/
+            /* if (System.IO.File.Exists(oldImagePath))
+             {
+                 System.IO.File.Delete(oldImagePath);
+             }*/
+
+            string productPath = @"image\products\product-" + id;
+            string finalPath = Path.Combine(_webHostEnvironment.WebRootPath, productPath);
+
+            if (!Directory.Exists(finalPath))
             {
-                System.IO.File.Delete(oldImagePath);
+                string[] filePaths = Directory.GetFiles(finalPath);
+                foreach (string filePath in filePaths)
+                {
+                    System.IO.File.Delete(filePath);
+                }
+
+                Directory.CreateDirectory(finalPath);
             }
+
+
+
             _unitOfWork.Product.Remove(productToBeDeleted);
             _unitOfWork.Save();
             
             return Json(new { success = true, message = "Delete successful" });
+        }
+
+        public IActionResult DeleteImage(int imageId)
+        {
+            var ImageTeBeDeleted = _unitOfWork.ProductImage.Get(u=>u.Id == imageId);
+            int productId = ImageTeBeDeleted.ProductId;
+
+            if(ImageTeBeDeleted != null)
+            {
+                if (!string.IsNullOrEmpty(ImageTeBeDeleted.ImageUrl))
+                {
+                    var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, ImageTeBeDeleted.ImageUrl.TrimStart('\\'));
+
+                    if (System.IO.File.Exists(oldImagePath))
+                    {
+                        System.IO.File.Delete(oldImagePath);
+                    }
+                }
+
+                _unitOfWork.ProductImage.Remove(ImageTeBeDeleted);
+                _unitOfWork.Save();
+
+                TempData["success"] = "Deleted successfully";
+            }
+
+            return RedirectToAction(nameof(Upsert), new { id = productId });
         }
 
         #endregion
